ready for smoke test

+++ scripts/nginx.bash
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [ -e /etc/nginx/sites-available/cropcompass ]
+  echo `date` "nginx already configured - exiting"
+  exit
+fi
+
 echo `date` "Configuring nginx"
 sudo cp /vagrant/scripts/cropcompass.nginx /etc/nginx/sites-available/cropcompass
 sudo ln -s /etc/nginx/sites-available/cropcompass /etc/nginx/sites-enabled
